Update for Down 4.0
We remove the custom wget implementation in favour of the new Down::Wget
backend. We also allow users to specify the Down backend to be used.

Since WebMock is still buggy (it's missing the #close method on its
StubSocket which Down attempts to call), we switch to making real HTTP
requests using httpbin.