Follow redirects in #exists?
It's possible that user will use a URL which redirects to the URL of the
actual file, so we modify #exists? to consider redirect as successful
response.

I first thought that we can just consider "3xx" as a successful response
and return true in case the HEAD request redirected. However, when URL
shorteners are used, #exists? would always return true. So we follow the
redirect.

Closes #2